build.sh: Set BUILD_DATETIME env variable.
BUILD_DATETIME ensures that both the Qssi and target
lunch images have the same date/timestamp properties.

This is particularly useful with incremental non-ab OTA where
fingerprint matching is done in recovery mode - recovery's
fingerprint vs the ota zip (system's/qssi's) fingerprint, and the
fingerprint has a field of date, which we match with this change.
Alternatively, BUILD_NUMBER can also be set, which also ensures
the fingerprints between recovery and Qssi (and ota.zip) match,
but this is generally done on oem side.

Change-Id: Ia290752cb34b01834b97d9c9897bbd502da7e1da